feat(esquema): las cantidades pasan a NUMERIC(14,4)
Nueve columnas de cuatro tablas: sale_items.quantity, las tres de stock,
las cuatro de stock_movements y pedido_items.cantidad. La base ya puede
representar 9,6.

Cuatro decimales y no tres porque recipe_items.quantity y
production_orders.quantity_produced ya son DECIMAL(12,4): con tres, una
produccion de 1,2345 se redondearia — el mismo defecto que esto viene a
eliminar, en espejo. Los tres decimales de una linea de venta los impone
un validador, no el tipo de columna.

El `up` saltea lo ya convertido y verifica SOLO lo que esa corrida
convirtio, asi que correrlo dos veces no aborta. El `down` se niega
condicionalmente: sin filas fraccionarias revierte limpio, con filas
fraccionarias tira nombrando tabla y cantidad. Va fuera de SE_NIEGAN
—el precedente es 20260804, no ese mapa— y hay una guardia que lo fija.

⚠ Corrige el modo de falla que la spec daba por bueno. El hallazgo
auditoria-frente2-hallazgos.json:335 dice que los consumos fraccionarios
«se redondean y el stock nunca baja». Medido: depende de como escribe
Sequelize. Un literal en el SQL —bulkCreate— redondea 9,6 a 10 en
silencio; un parametro bindeado —instance.update()— responde «invalid
input syntax for type integer». Produccion usa el segundo, asi que un
consumo fraccionario NO redondeaba: fallaba con 500 y revertia la orden
entera. La suposicion 3 de la spec queda corregida con la medicion.

Esa asimetria explica por que el defecto de la linea de venta en cero SI
era mudo: sale_items se escribe con bulkCreate.

Aparece ademas el punto de dibujo numero once, que el plan anoto como
riesgo y ninguna tabla listaba: GET /api/tiendanube/variantes.disponible,
que Tiendanube.jsx:858 dibuja. Se envuelve en el origen conservando el
null, que es lo que distingue «no hay fila de stock» de «hay cero»: un
Number(null) publicaria cero y agotaria una variante con mercaderia.

Y siete casos de cinco archivos de integracion ajenos —idempotencia de
ventas, los tres de TiendaNube, aislamiento y pedido publico— comparaban
contra numeros que ahora vuelven como texto. Se corrigen comparando
valores o contra el string que el contrato ya documenta.

⚠ NO DESPLEGAR SOLO: con el esquema migrado y sin el formateador de la
fase 4, el ticket imprime «3.0000 x Creatina». Las dos fases salen en la
misma release.

+++ apps/api/src/models/PedidoItem.js
@@ -31,7 +31,14 @@ const PedidoItem = sequelize.define('PedidoItem', {
   precio_lista: { type: DataTypes.DECIMAL(12, 2), allowNull: false },
   regla_id: { type: DataTypes.INTEGER, allowNull: true },
   // La EFECTIVAMENTE pedida, después del recorte por stock.
-  cantidad: { type: DataTypes.INTEGER, allowNull: false },
+  //
+  // DECIMAL(14,4) desde la migración 31, igual que las otras ocho columnas de
+  // cantidad. Hoy nada escribe decimales acá —`apps/tienda/src/carrito.js:55`
+  // hace `Math.floor` y la tienda pública sigue vendiendo por unidad—, pero el
+  // recorte por stock sale de `stock.quantity`, que sí puede ser fraccionario:
+  // dejarla en `INTEGER` sería sembrar el mismo redondeo silencioso en la única
+  // columna que quedó afuera. El tipo lo ata `tests/modeloPedidoItem.test.js`.
+  cantidad: { type: DataTypes.DECIMAL(14, 4), allowNull: false },
   subtotal: { type: DataTypes.DECIMAL(12, 2), allowNull: false },
 }, {
   tableName: 'pedido_items',
