+++ apps/api/src/models/StockMovement.js
@@ -27,20 +27,28 @@ const StockMovement = sequelize.define('StockMovement', {
     type: DataTypes.STRING(40),
     allowNull: true,
   },
+  // ── Las cuatro son DECIMAL(14,4) desde la migración 31 ──
+  //
+  // Son la foto de `stock.quantity` y `stock.available` antes y después de cada
+  // movimiento. Si la foto tuviera menos precisión que lo fotografiado, el
+  // historial diría que un stock pasó de 10 a 10 cuando pasó de 10 a 9,6: el
+  // único registro de qué se movió quedaría mintiendo, y en silencio.
+  //
+  // El tipo lo ata a la migración `tests/modeloStockMovement.test.js`.
   cantidad_anterior: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(14, 4),
     allowNull: false,
   },
   cantidad_nueva: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(14, 4),
     allowNull: false,
   },
   disponible_anterior: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(14, 4),
     allowNull: false,
   },
   disponible_nuevo: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(14, 4),
     allowNull: false,
   },
   usuario_id: {
